modified:   ../apps/UnitTwoDLib/FidTest.cpp 	modified:   ../apps/UnitTwoDLib/KDTreeTest.cpp 	modified:   ../apps/UnitTwoDLib/MeshAlgorithmTest.cpp 	modified:   ../apps/UnitTwoDLib/MeshTest.cpp 	modified:   ../apps/UnitTwoDLib/StatTest.cpp 	modified:   ../apps/UnitTwoDLib/TransitionMatrixGeneratorTest.cpp 	modified:   ../apps/UnitTwoDLib/TransitionMatrixTest.cpp 	modified:   ../libs/GeomLib/NumericalMasterEquationCode.hpp 	modified:   ../libs/GeomLib/SpikingOdeSystem.hpp 	modified:   ../libs/MPILib/include/MPINetworkCode.hpp 	modified:   ../libs/MPILib/src/utilities/ProgressBar.cpp 	modified:   ../libs/TwoDLib/CMakeLists.txt 	modified:   ../libs/TwoDLib/CSRMatrix.hpp 	modified:   ../libs/TwoDLib/MasterOMP.cpp 	modified:   ../libs/TwoDLib/MasterOMP.hpp 	modified:   ../libs/TwoDLib/MasterParameter.hpp 	modified:   ../libs/TwoDLib/MeshAlgorithm.hpp 	modified:   ../libs/TwoDLib/MeshAlgorithmCode.hpp 	modified:   ../libs/TwoDLib/Ode2DSystem.hpp 	modified:   ../libs/TwoDLib/TransitionMatrix.hpp 	modified:   ../libs/TwoDLib/TransitionMatrixGenerator.cpp 	modified:   ../python/visualize.py

+++ libs/GeomLib/SpikingOdeSystem.hpp
@@ -28,7 +28,7 @@ using MPILib::populist::ProbabilityQueue;
 
 namespace GeomLib {
 
-	//! In this system of ordinary differential equations
+	//! In this system of ordinary differential equations it is assumed that dynamics is always spiking.
 	class SpikingOdeSystem : public AbstractOdeSystem {
 	public:
 
@@ -42,10 +42,12 @@ namespace GeomLib {
 			const SpikingOdeSystem&
 		);
 
+		//! This is an abstract base class.
 		virtual ~SpikingOdeSystem() = 0;
 
-		void UpdateIndex() { if (--_index  == -_nr_bins ) _index = 0; }
+		void UpdateIndex() { if (--_index  == -static_cast<int>(_nr_bins) ) _index = 0; }
 
+		//! virtual construction by the framework
 		virtual SpikingOdeSystem* Clone() const = 0;
 
 		virtual Potential DCContribution() const = 0;
